Add a simple I2C wrapper to replace WiringPi
WiringPi is a fairly heavyweight wrapper to keep a standard easy interface to a whole slew of devices and interfaces, adding a few layers of un-necessary abstraction. GBCS is hitting CPU limits on the Pi where it can't put data onto the bus fast enough without overclocking. Adding this class should remove the WiringPi dependency and provide the minimum data path to the I2C bus without going to kernel mode